+++ lib/animations/size_animation_transition.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:page_animation_transition/page_animation_interface.dart';
+
+class SizeAnimationTransition implements PageAnimationInterface {
+  @override
+  Widget animate(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return Align(
+      child: SizeTransition(
+        sizeFactor: animation,
+        child: child,
+      ),
+    );
+  }
+}